requirements.txt
fastapi[standard]
redis==6.4.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
python-dotenv==1.1.1